atan: optimize 20% via `flip_sign_nonnan`
The flip_sign_nonnan function will xor the sign of one input onto
another float, which allows for implementing `signum(x) * y`:
`flip_sign_nonnan(x, y)` as expected, as long as `x` is non-nan. This
operation occurs quite a few times in the various atan calculations.

`atan` is rearranged so that it can use this, by ensuring `x` won't be
NaN on the path that needs it, which gives a nice speed-up:

```
atan/full               time:   [25.317 ns 25.620 ns 25.888 ns]
                        change: [-23.188% -22.167% -21.123%] (p = 0.00 < 0.05)
```