Implement log2_1p.
This switches between the linear approximation log_B(x) ~ x/ln(B) for
small x and the "obvious" log_B(1 + x), with the size threshold chosen
by `optimizer`. There was not much value in optimizing the
multiplicative constant (especially true for `ln_1p` if that is every
implemented: ln(x) ~ x avoids a multiplication), or using a
higher-degree polynomial, because the equation above is as accurate as
the `log2` approximation.

Fixes #11.